Add own vertex rendering

+++ App/src/main.cpp
@@ -31,57 +31,59 @@ int main(int argc, char** argv)
 
     // GM: Will make a tests file, just doing this for now.
     // Test start.
-
-    Velox::EntityHandle handle = entityManager.createEntity();
-    Velox::Entity* newEntity = entityManager.getMut(handle);
-
-    printf("before modification: %i\n", newEntity->test);
-
-    newEntity->position.x += 1;
-    newEntity->test += 1;
-
-    printf("after modification: %i\n", newEntity->test);
-
-    newEntity = entityManager.getMut(handle);
-
-    printf("after modification (re-query): %i\n", newEntity->test);
-
-    printf("\n");
-    for (auto [handle, entity] : entityManager.iter())
+    if (false)
     {
-        printf("Iterator test 1: %i\n", entity->test);
+        Velox::EntityHandle handle = entityManager.createEntity();
+        Velox::Entity* newEntity = entityManager.getMut(handle);
+
+        printf("before modification: %i\n", newEntity->test);
+
+        newEntity->position.x += 1;
+        newEntity->test += 1;
+
+        printf("after modification: %i\n", newEntity->test);
+
+        newEntity = entityManager.getMut(handle);
+
+        printf("after modification (re-query): %i\n", newEntity->test);
+
+        printf("\n");
+        for (auto [handle, entity] : entityManager.iter())
+        {
+            printf("Iterator test 1: %i\n", entity->test);
+        }
+
+        entityManager.destroyEntity(handle);
+
+        entityManager.createEntity();
+        entityManager.createEntity();
+
+        printf("\n");
+        for (auto [handle, entity] : entityManager.iter())
+        {
+            printf("Iterator test 2: %i\n", entity->test);
+        }
+
+        Velox::EntityHandle newHandle = entityManager.createEntity();
+        newEntity = entityManager.getMut(newHandle);
+
+        entityManager.createEntity();
+
+        printf("\n");
+        for (auto [handle, entity] : entityManager.iter())
+        {
+            printf("Iterator test 3: %i\n", entity->test);
+        }
+
+        newEntity->test += 1;
+
+        printf("\n");
+        for (auto [handle, entity] : entityManager.iter())
+        {
+            printf("Iterator test 4: %i\n", entity->test);
+        }
+        // Test end.
     }
-
-    entityManager.destroyEntity(handle);
-
-    entityManager.createEntity();
-    entityManager.createEntity();
-
-    printf("\n");
-    for (auto [handle, entity] : entityManager.iter())
-    {
-        printf("Iterator test 2: %i\n", entity->test);
-    }
-
-    Velox::EntityHandle newHandle = entityManager.createEntity();
-    newEntity = entityManager.getMut(newHandle);
-
-    entityManager.createEntity();
-
-    printf("\n");
-    for (auto [handle, entity] : entityManager.iter())
-    {
-        printf("Iterator test 3: %i\n", entity->test);
-    }
-
-    newEntity->test += 1;
-
-    printf("\n");
-    for (auto [handle, entity] : entityManager.iter())
-    {
-        printf("Iterator test 4: %i\n", entity->test);
-    }
-    // Test end.
 
     while (!Velox::QuitRequested())
     {
